Fix nav links on mobile
On mobile, the screen is smaller so the links in the header wrap around. This causes the nav to be much taller than 100px, but since the height is hardcoded at 100px, this causes the lower nav links on the screen to be visible but not clickable as they are covered up by the main page div. This PR changes `height` to `min-height` so that when the height of the nav grows due to text wrapping on mobile, everything still works as expected and the height of the nav can grow to accommodate the links.